# Add/get block name from path convention