--- a/projects/packages/blocks/changelog/add-get_block_name_from_path_convention
+++ b/projects/packages/blocks/changelog/add-get_block_name_from_path_convention
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Blocks: Determine block names from filename convention instead of disk access